requirements.txt
faiss-cpu
numpy
pypdf
sentence-transformers
tiktoken
python-dotenv

